Serve statics through Rocket in the server integration example
When using Rocket as an http server, it seems logical to also use it to
serve the index page and the wasm/js package files. This allows us to get
rid of rocket_cors and it simplifies the server code quite a bit.

This commit also bumps rocket to 0.4.0 and imrpoves the README.